Changed type of id fields from 'byte' to 'int' to be compatible with Entity Framework.

+++ src/DND_Spells/SpellsDomain/Model/Area.cs
@@ -7,7 +7,7 @@ public partial class Area : Entity
 {
     public int AreaId { get; set; }
 
-    public byte AreaTypeId { get; set; }
+    public int AreaTypeId { get; set; }
 
     public int AreaSize { get; set; }
