+++ src/DND_Spells/SpellsDomain/Model/Effect.cs
@@ -7,7 +7,7 @@ public partial class Effect : Entity
 {
     public int SpellId { get; set; }
 
-    public byte EffectTypeId { get; set; }
+    public int EffectTypeId { get; set; }
 
     public virtual EffectType EffectType { get; set; } = null!;
