+++ src/DND_Spells/SpellsDomain/Model/EffectType.cs
@@ -5,7 +5,7 @@ namespace SpellsDomain.Model;
 
 public partial class EffectType : Entity
 {
-    public byte EffectTypeId { get; set; }
+    public int EffectTypeId { get; set; }
 
     public string Label { get; set; } = null!;
 }
